功能：模块化部署系统并新增每日一句功能 将原有的单体式 deployment.py 重构为多个职责清晰的模块化核心部署器（BaseDeployer、VersionManager、MaiBotDeployer、MoFoxBotDeployer、NapCatDeployer），实现关注点分离与版本管理 更新 napcat_downloader，使其委托给对应的 NapCatDeployer 实现，并添加缓存清理辅助函数 新增每日一句数据文件（data/Golden_sentence.json），并在 main_refactored.py、ui.menus.py 和 ui.theme.py 中加入刷新支持 保持向后兼容性：保留 deployment_old.py 备份，且 deployment_manager 外观接口仍可正常使用 无破坏性变更
