+++ src/modules/deployment_core/__init__.py
@@ -1,0 +1,19 @@
+# -*- coding: utf-8 -*-
+"""
+部署核心模块
+提供模块化的部署功能
+"""
+
+from .base_deployer import BaseDeployer
+from .version_manager import VersionManager
+from .maibot_deployer import MaiBotDeployer
+from .mofox_deployer import MoFoxBotDeployer
+from .napcat_deployer import NapCatDeployer
+
+__all__ = [
+    'BaseDeployer',
+    'VersionManager',
+    'MaiBotDeployer',
+    'MoFoxBotDeployer',
+    'NapCatDeployer',
+]
